clean up code for latest compare features
